Refactor makeSongQuery function
