+++ vanguarde/src/redux/common.js
@@ -1,4 +1,6 @@
 import {store} from 'react-notifications-component';
+import {fetchBegin, fetchError, fetchSongListSuccess} from './songs';
+import axios from 'axios';
 
 export function handleSuccess(response) {
     store.addNotification({
@@ -33,3 +35,21 @@ export function handleError(errorObj) {
     });
 }
 
+export function makeSongQuery(dispatch, getState, {apiConfig}, searchQuery = {}) {
+    dispatch(fetchBegin());
+    console.log({searchQuery});
+    const searchState = getState().searchReducer;
+    const searchParams = {
+        'name': searchState.searchQuery ? searchState.searchQuery : null,
+        'sort_by': searchState.sorting ? 'name|' + searchState.sorting : null,
+    };
+    axios.get(`${apiConfig.url}songs`, {params: searchParams})
+        .then((res) => {
+            dispatch(fetchSongListSuccess(res.data));
+            handleSuccess(res);
+        })
+        .catch((error) => {
+            dispatch(fetchError(error));
+            handleError(error);
+        });
+}
